Final commit

+++ OS/HW3/find.h
@@ -7,7 +7,7 @@
  // FOR FINDING THE DESIRED DIRECTORY WITHIN THE WORKING DIRECTORY (RELATIVE) OR IN THE FILE SYSTEM (ABSOLUTE)
  int32_t findDirectory(FILE* disk, int32_t cluster, BPB_struct & boot_sector, vector<string> dir);
 
-uint16_t findEmptyCluster(FILE* disk,BPB_struct boot_sector);
+int findEmptyCluster(FILE* disk,BPB_struct boot_sector);
 
 int32_t findFile(FILE* disk, int32_t cluster, BPB_struct boot_sector, vector<string> path);
